// components/create/types.ts
export interface SocialHandles {
  twitter: string;
  linkedin: string;
  website: string;
  email: string;
  youtube: string;
  buymeacoffee: string;
  instagram: string;
  facebook: string;
  twitch: string;
}

export const defaultSocialHandles: SocialHandles = {
  twitter: "",
  linkedin: "",
  website: "",
  email: "",
  youtube: "",
  instagram: "",
  buymeacoffee: "",
  facebook: "",
  twitch: "",
};